% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{flatten_with_id}
\alias{flatten_with_id}
\title{Utility function to flatten nested objects into tabular format}
\usage{
flatten_with_id(x, id, id_col = "Diarienummer")
}
\arguments{
\item{x}{the object to flatten}

\item{id}{vector of corresponding identifiers}

\item{id_col}{string with title for identifier column, Default: 'Diarienummer'}
}
\value{
table with flattened results
}
\description{
Utility function to flatten nested objects into tabular format
}
