% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vinnova}
\alias{vinnova}
\title{Data for entities such as programmes, call for proposals, rounds or funded projects}
\usage{
vinnova(
  kind = c("programmes", "proposals", "rounds", "projects"),
  from_date,
  id
)
}
\arguments{
\item{kind}{the type of entity, Default: c("programmes", "proposals", "rounds", "projects")}

\item{from_date}{the date since when changes should be requested, if not set, since three days back}

\item{id}{identifier (or comma separated identifiers) for specific entities}
}
\value{
table with results
}
\description{
This function take the kind of entity requested and also either accepts a vector of identifiers or a date since which changes should be returned
}
\examples{
\dontrun{
if(interactive()){
 vinnova("programmes", from_date = Sys.Date())
 }
}
}
