% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vinnova_file}
\alias{vinnova_file}
\title{Files or document contents from Vinnova documents}
\usage{
vinnova_file(id)
}
\arguments{
\item{id}{identifier (use DocumentID field)}
}
\value{
raw binary file content
}
\description{
Files or document contents from Vinnova documents
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
