% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vinnova_files}
\alias{vinnova_files}
\title{Files for latest vinnova data in tabular format}
\usage{
vinnova_files(changes = vinnova_latest())
}
\arguments{
\item{changes}{results from vinnova_latest fcn, by default}
}
\value{
table with blob data for results including file contents
}
\description{
Files for latest vinnova data in tabular format
}
