% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vinnova_latest}
\alias{vinnova_latest}
\title{Flattened complete data for calls, programmes, rounds and projects since a given date}
\usage{
vinnova_latest(from_date = Sys.Date() - 3)
}
\arguments{
\item{from_date}{the date since changes are requested, Default: Sys.Date() - 3}
}
\value{
object with tables
}
\description{
Flattened complete data for calls, programmes, rounds and projects since a given date
}
\examples{
\dontrun{
if(interactive()){
 vinnova_latest()
 }
}
}
